+++ tools/xdhq/proxy/LIB/slfhlead.cpp
@@ -17,24 +17,39 @@
 	along with 'XDHq'.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#ifndef COMMON_INC_
-# define COMMON_INC_
+#include "slfhlead.h"
 
-# include "xdhcuc.h"
+using namespace slfhlead;
 
-# include "csdmnc.h"
-# include "logq.h"
-
-namespace common {
-	namespace faas {
-		using namespace xdhcmn::faas;
-	}
-
-	extern logq::rFDriver<> LogDriver;
-
-	void SetCallback(xdhcuc::cGlobal &Callback);
-
-	xdhcuc::cGlobal &GetCallback(void);
+namespace {
+	csdmnc::rCore Core_;
+	bso::sBool CoreIsInitialized_ = false;
 }
 
-#endif
+bso::sBool slfhlead::InitializeCore(
+		const char *HostService,
+		bso::uint__ PingDelay,
+		sck::duration__ Timeout,
+		csdmnc::cLog *LogCallback)
+{
+		if ( CoreIsInitialized_ )
+			qRFwk();
+
+		if ( Core_.Init(HostService, PingDelay, Timeout, LogCallback) ) {
+			CoreIsInitialized_ = true;
+			return true;
+		} else
+			return false;
+}
+
+bso::sBool slfhlead::CoreIsInitialized(void)
+{
+	return CoreIsInitialized_;
+}
+
+csdmnc::rCore &slfhlead::Core(void) {
+	if ( !CoreIsInitialized_ )
+		qRGnr();
+
+	return Core_;
+}