+++ tools/xdhq/proxy/LIB/xdhqxdh.cpp
@@ -22,6 +22,7 @@
 #include "common.h"
 #include "registry.h"
 #include "session.h"
+#include "slfhlead.h"
 
 #include "idsq.h"
 
@@ -57,7 +58,7 @@ namespace {
 				Translation.Init();
 				sclm::GetBaseTranslation("ConnectingTo", Translation, HostService);
 				cio::COut << Translation << txf::commit << txf::nl;
-				if ( common::InitializeCore( HostService.Convert( Buffer ), 0, sck::NoTimeout ) ) {
+				if ( slfhlead::InitializeCore( HostService.Convert( Buffer ), 0, sck::NoTimeout ) ) {
 					Translation.Init();
 					sclm::GetBaseTranslation("ConnectedTo", Translation, HostService);
 					cio::COut << Translation << txf::commit << txf::nl;