+++ tools/xdhq/wrappers/JRE/DOM_SHRD.java
@@ -39,6 +39,4 @@ public abstract class DOM_SHRD {
 	public abstract void getAction(Event event);
 
 	public abstract Object call(String command, Type type, info.q37.xdhq.ARG ...args);
-
-	public abstract boolean isQuitting();
 }